feat: implementing model for the Administrator class
