+++ src/com/br/model/DAO/AdministratorsDAO.java
@@ -1,0 +1,7 @@
+package com.br.model.DAO;
+
+import com.br.model.entity.Administrators;
+
+public class AdministratorsDAO extends BaseDAO<Administrators> {
+
+}
