+++ src/com/br/model/entity/Administrators.java
@@ -10,82 +10,6 @@ public class Administrators {
   private String password;
   private String accessCode;
 
-  // simulate a database
-  public static List<Administrators> fakeBD = new ArrayList<Administrators>();
-
-  // implement the constructors 
-  public Administrators(
-    String name, 
-    String user, 
-    String password, 
-    String code
-  ) {
-    this.setName(name);
-    this.setUser(user);
-    this.setPassword(password);
-    this.setAccessCode(code);
-  }
-
-  public Administrators() {}
-
-  public boolean login(String user, String password) {
-    // implements a fake login while not database yet and the tools that will be used
-    if (user.isEmpty() || password.isEmpty()) {
-      System.out.println("Dados vazios! Tente novamente");
-      return false;
-    }
-
-    for (Administrators admin : fakeBD) {
-      if (admin.getUser().equals(user) && admin.getPassword().equals(password)) {
-        System.out.println("Logado com sucesso!");
-        return true;
-      }
-    }
-
-    System.out.println("Usuário ou Senha incorreto!");
-    return false;
-  }
-
-  public void logout() {
-    // implements a fake logout while not active login session
-    System.out.println("Logout");
-  }
-
-  public void registerAdministrator(Administrators newAdmin) {
-    // implements a fake register while not a database
-    if (
-      newAdmin.getName().isEmpty() || 
-      newAdmin.getUser().isEmpty() ||
-      newAdmin.getPassword().isEmpty() || 
-      newAdmin.getAccessCode().isEmpty()
-    ) {
-      System.out.println("Dados não podem estarem vazios!");
-      return;
-    } 
-
-    for (Administrators admin : fakeBD) {
-      if (
-        admin.getAccessCode().equals(newAdmin.getAccessCode()) ||
-        admin.getUser().equals(newAdmin.getUser())
-      ) {
-        System.out.println("Código de acesso ou usuário já existente!");
-        return;
-      }
-    }
-
-    fakeBD.add(newAdmin);
-  }
-
-  // implements reporting functions
-  // but the logic is still a little iffy
-  public void generateRentsByCustomerReport() {}
-
-  public void generateReportAllRented() {}
-
-  public void generateMonthlyBillingReport() {
-
-  }
-
   // implements GETTERS and SETTERS to ensure encapsulation of attributes
   public String getName() {
     return this.name;
